Modify NavigationIteratorBase to (almost) conform to std forward iterator requirements
The `const bool` in `VertexNeighborIteratorState` was causing the default constructor for `NavigationIteratorBase<VertexAdjacentVertexNavigator>` and friends by implicitly deleting default constructors up the chain. This was solved by making the member variables private and adding `const` accessors. The functionality of the class should be the same, just a few renamings to avoid clashes