Bump Eigen version to 3.3.9 (#225)
Eigen 3.3.8 is known to have a bug. Somehow geometry-central and its tests never trigger the problematic code, but is sometimes triggered when geometry-central is used within other projects (and only on Linux, but not on Mac -- I guess the codepath is different on macOS vs. linux). The bug is fixed in the next version (3.3.9), so fetch that version of Eigen instead.